Move trait Fill to rand_core
