+++ src/rng.rs
@@ -11,9 +11,7 @@
 
 use crate::distr::uniform::{SampleRange, SampleUniform};
 use crate::distr::{self, Distribution, StandardUniform};
-use core::num::Wrapping;
-use core::{mem, slice};
-use rand_core::RngCore;
+use rand_core::{Fill, RngCore};
 
 /// User-level interface for RNGs
 ///
@@ -356,107 +354,13 @@ pub trait Rng: RngCore {
 
 impl<R: RngCore + ?Sized> Rng for R {}
 
-/// Support filling a slice with random data
-///
-/// This trait allows slices of "plain data" types to be efficiently filled
-/// with random data.
-///
-/// Implementations are expected to be portable across machines unless
-/// clearly documented otherwise (see the
-/// [Chapter on Portability](https://rust-random.github.io/book/portability.html)).
-/// The implementations provided achieve this by byte-swapping on big-endian
-/// machines.
-pub trait Fill: Sized {
-    /// Fill this with random data
-    fn fill<R: Rng + ?Sized>(this: &mut [Self], rng: &mut R);
-}
-
-impl Fill for u8 {
-    fn fill<R: Rng + ?Sized>(this: &mut [Self], rng: &mut R) {
-        rng.fill_bytes(this)
-    }
-}
-
-/// Call target for unsafe macros
-const unsafe fn __unsafe() {}
-
-/// Implement `Fill` for given type `$t`.
-///
-/// # Safety
-/// All bit patterns of `[u8; size_of::<$t>()]` must represent values of `$t`.
-macro_rules! impl_fill {
-    () => {};
-    ($t:ty) => {{
-        // Force caller to wrap with an `unsafe` block
-        __unsafe();
-
-        impl Fill for $t {
-            fn fill<R: Rng + ?Sized>(this: &mut [Self], rng: &mut R) {
-                if this.len() > 0 {
-                    let size = mem::size_of_val(this);
-                    rng.fill_bytes(
-                        // SAFETY: `this` non-null and valid for reads and writes within its `size`
-                        // bytes. `this` meets the alignment requirements of `&mut [u8]`.
-                        // The contents of `this` are initialized. Both `[u8]` and `[$t]` are valid
-                        // for all bit-patterns of their contents (note that the SAFETY requirement
-                        // on callers of this macro). `this` is not borrowed.
-                        unsafe {
-                            slice::from_raw_parts_mut(this.as_mut_ptr()
-                                as *mut u8,
-                                size
-                            )
-                        }
-                    );
-                    for x in this {
-                        *x = x.to_le();
-                    }
-                }
-            }
-        }
-
-        impl Fill for Wrapping<$t> {
-            fn fill<R: Rng + ?Sized>(this: &mut [Self], rng: &mut R) {
-                if this.len() > 0 {
-                    let size = this.len() * mem::size_of::<$t>();
-                    rng.fill_bytes(
-                        // SAFETY: `this` non-null and valid for reads and writes within its `size`
-                        // bytes. `this` meets the alignment requirements of `&mut [u8]`.
-                        // The contents of `this` are initialized. Both `[u8]` and `[$t]` are valid
-                        // for all bit-patterns of their contents (note that the SAFETY requirement
-                        // on callers of this macro). `this` is not borrowed.
-                        unsafe {
-                            slice::from_raw_parts_mut(this.as_mut_ptr()
-                                as *mut u8,
-                                size
-                            )
-                        }
-                    );
-                    for x in this {
-                        *x = Wrapping(x.0.to_le());
-                    }
-                }
-            }
-        }}
-    };
-    ($t:ty, $($tt:ty,)*) => {{
-        impl_fill!($t);
-        // TODO: this could replace above impl once Rust #32463 is fixed
-        // impl_fill!(Wrapping<$t>);
-        impl_fill!($($tt,)*);
-    }}
-}
-
-// SAFETY: All bit patterns of `[u8; size_of::<$t>()]` represent values of `u*`.
-const _: () = unsafe { impl_fill!(u16, u32, u64, u128,) };
-// SAFETY: All bit patterns of `[u8; size_of::<$t>()]` represent values of `i*`.
-const _: () = unsafe { impl_fill!(i8, i16, i32, i64, i128,) };
-
 #[cfg(test)]
 mod test {
     use super::*;
     use crate::test::{const_rng, rng};
     #[cfg(feature = "alloc")]
     use alloc::boxed::Box;
+    use core::num::Wrapping;
 
     #[test]
     fn test_fill_bytes_default() {
